practice_007a_opentelemetry/pyproject.toml
[project]
name = "practice-007a-opentelemetry"
version = "0.1.0"
description = "OpenTelemetry instrumentation practice: manual tracing, context propagation, Jaeger"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.28.0",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.29.0",
    "uvicorn>=0.34.0",
]
